:art: refactor: UNIC-1785 Improve task creation + Create publish excel task
